Creating a file for ConfigurationMixin and helper functions
